First Part complite

+++ blogApp/views.py
@@ -1,5 +1,8 @@
+from urllib.parse import quote_plus
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.utils import timezone
+from django.db.models import Q
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -9,11 +12,15 @@ from django.core.paginator import Paginator
 
 
 def post_create(request):
-    form = PostForm(request.POST or None)
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
+    form = PostForm(request.POST or None, request.FILES or None)
 
     if form.is_valid():
         instance = form.save(commit=False)
         # form.cleaned_data.get("")
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully create")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -32,26 +39,49 @@ def post_create(request):
 
 
 def post_detail(request, id):
-    instance = get_object_or_404(Post, id=id)
+    instance        = get_object_or_404(Post, id=id)
+
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
     # instance = Post.objects.get(id = 1)
+    share_string    = quote_plus(instance.content)
+
     context = {
         "title": instance.title,
         "instance": instance,
+        "share_string": share_string,
     }
     return render(request, "post_detail.html", context)
 
 
 def post_list(request):
-    queryset_list = Post.objects.all()  # .order_by("-timestamp")
-    paginator = Paginator(queryset_list, 10) # Show 25 contacts per page
+    today = timezone.now().date()
+    queryset_list    = Post.objects.active()  # .order_by("-timestamp") # filter(draft=False).filter(publish__lte=timezone.now())
+
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Post.objects.all()
+
+    query            = request.GET.get("q")
+
+    if query:
+        queryset_list = queryset_list.filter(
+            Q(title__icontains=query)|
+            Q(content__icontains=query)|
+            Q(user__first_name__icontains=query)|
+            Q(user__last_name__icontains=query)
+        ).distinct()
+
+    paginator        = Paginator(queryset_list, 5) # Show 25 contacts per page
     page_request_var = "page"
-    page = request.GET.get(page_request_var)
-    queryset = paginator.get_page(page)
+    page             = request.GET.get(page_request_var)
+    queryset         = paginator.get_page(page)
 
     context = {
         "Title": "Post List",
         "object_list": queryset,
         "page_request_var": page_request_var,
+        "today" : today
     }
 
     # if request.user.is_authenticated:
@@ -66,8 +96,12 @@ def post_list(request):
 
 
 def post_update(request, id=None):
-    instance = get_object_or_404(Post, id=id)
-    form = PostForm(request.POST or None, instance=instance)
+
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
+    instance    = get_object_or_404(Post, id=id)
+    form        = PostForm(request.POST or None, request.FILES or None, instance=instance)
 
     if form.is_valid():
         instance = form.save(commit=False)
@@ -88,6 +122,10 @@ def post_update(request, id=None):
 
 
 def post_delete(request, id=None):
+
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, "Successfully Delete")
